Remove legacy model
