+++ src/dotnet/APIView/APIView/Model/CodeFile.cs
