+++ src/dotnet/APIView/APIView/Model/NavigationItem.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+
+namespace APIView
+{
+    public class NavigationItem
+    {
+        public string Text { get; set; }
+        public string NavigationId { get; set; }
+        public NavigationItem[] ChildItems { get; set; } = Array.Empty<NavigationItem>();
+
+        public Dictionary<string, string> Tags { get; set; } = new Dictionary<string, string>(0);
+
+        public override string ToString() => Text;
+    }
+}
