+++ src/dotnet/APIView/APIView/Program.cs
@@ -8,7 +8,7 @@ namespace ApiView
         {
             try
             {
-                var assemblySymbol = AssemblyApiView.GetCompilation(args[0]);
+                var assemblySymbol = CompilationFactory.GetCompilation(args[0]);
                 var renderer = new CodeFileRenderer();
                 var codeNode = new CodeFileBuilder().Build(assemblySymbol, false);
                 Console.WriteLine(renderer.Render(codeNode));
