--- src/dotnet/APIView/APIViewTest/AssemblyTests.cs
@@ -1,35 +1,0 @@
-using Microsoft.CodeAnalysis;
-using ApiView;
-using Xunit;
-
-namespace APIViewTest
-{
-    public class NodeBuilderTests
-    {
-
-    }
-
-    public class AssemblyTests
-    {
-        [Fact]
-        public void AssemblyTestAssemblyFromSymbol()
-        {
-            IAssemblySymbol assemblySymbol = TestResource.GetAssemblySymbol();
-            AssemblyApiView assembly = new AssemblyApiView(assemblySymbol);
-
-            Assert.Equal("TestLibrary", assembly.Name);
-
-            Assert.Single(assembly.GlobalNamespace.Namespaces);
-        }
-
-        [Fact]
-        public void AssemblyTestAssembliesFromFile()
-        {
-            AssemblyApiView assembly = AssemblyApiView.AssemblyFromFile("TestLibrary.dll");
-            Assert.Equal("TestLibrary", assembly.Name);
-
-            NamespaceApiView globalNamespace = assembly.GlobalNamespace;
-            Assert.Single(globalNamespace.Namespaces);
-        }
-    }
-}
