+++ src/dotnet/APIView/APIViewWeb/Pages/Assemblies/LegacyReview.cshtml
@@ -1,0 +1,14 @@
+﻿@page
+
+@model APIViewWeb.Pages.Assemblies.LegacyReview
+@{
+    ViewData["Title"] = "Legacy Review";
+}
+
+<h2>This review can not be viewed anymore but here are the comments:</h2>
+
+@foreach (var commentThread in Model.Comments)
+{
+    <pre>@commentThread.Key</pre>
+    <partial name="_CommentThreadPartial" model="@new APIViewWeb.Models.CommentThreadModel() { AssemblyId = Model.Id, Comments = commentThread.Value, LineId = commentThread.Key }" />
+}
